pyproject.toml
[tool.poetry]
name = "segment-data-generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
analytics-python = "^1.4.0"
Faker = "^13.3.4"
faker-commerce = "^1.0.3"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
